+++ src/routes/user.ts
@@ -1,10 +1,9 @@
 import { Router } from 'express'
 
 import { createUser } from '@controllers/User/createUser'
-import { getUser } from '@controllers/User/getUser'
-import { getUsers } from '@controllers/User/getUsers'
 import { deleteUser } from '@controllers/User/deleteUser'
 import { putUser } from '@controllers/User/putUser'
+import { getUser, getUsers } from '@controllers/User/readUser'
 
 const userRouter = Router()
